feat: implement HistoryScreen with order history display, including loading and error states

+++ shared/src/commonMain/kotlin/com/cashierserviceapp/domain/models/Order.kt
@@ -2,6 +2,12 @@ package com.cashierserviceapp.domain.models
 
 import kotlinx.serialization.Serializable
 
+/**
+ * `OrderSummaryDTO` — the row shape shared by the in-progress and history lists.
+ *
+ * [createdAt] arrives as an ISO-8601 instant in UTC (`2026-08-13T17:16:47Z`) and is kept as text
+ * here; the screens that show it parse and localise it themselves.
+ */
 @Serializable
 data class Order(
     val id: String,
@@ -9,7 +15,8 @@ data class Order(
     val customerName: String,
     val orderCode: String,
     val createdAt: String,
-    val totalCost: Int,
+    // Int64 server-side, so Long here: rupiah totals outgrow Int sooner than you'd think.
+    val totalCost: Long,
     val itemsCount: Int,
 )
